requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
loguru==0.6.0
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
schedule==1.1.0
selenium==4.4.3
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
telebot==0.0.4
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.11
websockets==10.3
wsproto==1.1.0
xvfbwrapper==0.2.9
